create preview (not subitem) for ufed chats, go to message chat position
