+++ iped-app/resources/config/conf/CustomSignatures.xml
@@ -923,8 +923,12 @@
     	<sub-class-of type="text/html"/>
     </mime-type>
     
-    <mime-type type="application/x-ufed-chat-preview">
+    <mime-type type="application/x-ufed-chat">
     	<sub-class-of type="application/x-preview-with-links"/>
+    </mime-type>
+    
+    <mime-type type="application/x-ufed-chat-whatsapp">
+    	<sub-class-of type="application/x-ufed-chat"/>
     </mime-type>
     
     <mime-type type="application/x-win10-timeline">
