+++ iped-app/resources/config/profiles/pt-BR/linkanalysis/conf/MakePreviewConfig.txt
@@ -10,3 +10,4 @@ supportedMimes = application/x-prefetch
 
 # List of mimetypes which parsers insert links to other case items into preview
 supportedMimesWithLinks = application/x-emule; application/x-ares-galaxy; application/x-shareaza-library-dat
+supportedMimesWithLinks = application/x-ufed-chat; application/x-ufed-chat-whatsapp
