+++ iped-app/src/main/java/dpf/sp/gpinf/indexer/desktop/ResultTableModel.java
@@ -258,8 +258,13 @@ public class ResultTableModel extends AbstractTableModel implements SearchResult
             StringBuilder sb = new StringBuilder();
             for (int i = 0; i < values.length; i++) {
                 sb.append(values[i]);
-                if (i != values.length - 1)
+                if (i != values.length - 1) {
+                    if(i == 9) {
+                        sb.append(" ..."); //$NON-NLS-1$
+                        break;
+                    }
                     sb.append(" | "); //$NON-NLS-1$
+                }
             }
 
             value = sb.toString().trim();
