+++ iped-engine/src/main/java/dpf/sp/gpinf/indexer/process/P2PBookmarker.java
@@ -57,7 +57,9 @@ public class P2PBookmarker {
                 new P2PProgram(HashTask.HASH.MD5.toString(), "Shareaza")); //$NON-NLS-1$
         p2pPrograms.put(WhatsAppParser.WHATSAPP_CHAT.toString(),
                 new P2PProgram(HashTask.HASH.SHA256.toString(), "WhatsApp")); //$NON-NLS-1$
-        p2pPrograms.put(UFEDChatParser.UFED_CHAT_PREVIEW_MIME.toString(),
+        p2pPrograms.put(UFEDChatParser.UFED_CHAT_MIME.toString(),
+                new P2PProgram(IndexItem.HASH.toString(), "UFED_Chats")); //$NON-NLS-1$
+        p2pPrograms.put(UFEDChatParser.UFED_CHAT_WA_MIME.toString(),
                 new P2PProgram(IndexItem.HASH.toString(), "UFED_Chats")); //$NON-NLS-1$
         p2pPrograms.put(SkypeParser.FILETRANSFER_MIME_TYPE, new P2PProgram(IndexItem.HASH, "Skype")); //$NON-NLS-1$
         p2pPrograms.put(SkypeParser.CONVERSATION_MIME_TYPE, new P2PProgram(IndexItem.HASH, "Skype")); //$NON-NLS-1$
@@ -102,7 +104,8 @@ public class P2PBookmarker {
                     continue;
 
                 String bookmarkSufix = program.appName;
-                if (UFEDChatParser.UFED_CHAT_PREVIEW_MIME.toString().equals(mediaType)) {
+                if (UFEDChatParser.UFED_CHAT_MIME.toString().equals(mediaType) ||
+                        UFEDChatParser.UFED_CHAT_WA_MIME.toString().equals(mediaType)) {
                     String source = doc.get(ExtraProperties.UFED_META_PREFIX + "Source"); //$NON-NLS-1$
                     if (source != null)
                         bookmarkSufix = source;
